refactor(server): extract shared paseo tool registry + browser tools catalog
Pull the tool-registry primitive (registerTool/toCatalog/parseToolInput) out of
createPaseoToolCatalog into a reusable createPaseoToolRegistry, and add
createBrowserToolsCatalog so a non-MCP caller can build the same browser tool
catalog the MCP server exposes. No behaviour change for existing callers.
